requirements.txt
openai
python-dotenv
notebook
kor
langchain
langchain-together
rich
requests
twilio
uvicorn
fastapi
pydantic
prisma
python-multipart
pillow
matplotlib
uuid